feat: Move loading to dedicated server class and break it down in smaller chunks
